adjust file struct
